requirements.txt
# pip install -r requirements.txt
pillow
requests
jupyter
numpy
pandas
matplotlib
seaborn
graphviz
Cython
ultralytics
opencv-python>=4.1.2
tensorflow==1.9.0
keras==2.16
pyYAML
tqdm